Added postgresql support and application.properties optionss, renamed NetworkScan(Service/Scheduler) to Scan(Service/Scheduler), some nicer changes
