+++ backend/src/main/java/com/danny/lansentinel/controller/DeviceController.java
@@ -2,14 +2,13 @@ package com.danny.lansentinel.controller;
 
 import com.danny.lansentinel.entity.Device;
 import com.danny.lansentinel.repository.DeviceRepository;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("/devices")
+@RequestMapping("/device")
 public class DeviceController {
 
     private final DeviceRepository deviceRepository;
