+++ backend/src/main/java/com/danny/lansentinel/controller/ScanController.java
@@ -1,7 +1,7 @@
 package com.danny.lansentinel.controller;
 
-import com.danny.lansentinel.scheduler.NetworkScanScheduler;
-import com.danny.lansentinel.service.NetworkScannerService;
+import com.danny.lansentinel.scheduler.ScanScheduler;
+import com.danny.lansentinel.service.ScanService;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -10,13 +10,13 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/scan")
 public class ScanController {
 
-    NetworkScannerService networkScannerService;
-    NetworkScanScheduler networkScanScheduler;
+    private final ScanService scanService;
+    private final ScanScheduler scanScheduler;
 
-    public ScanController(NetworkScannerService networkScannerService,
-                          NetworkScanScheduler networkScanScheduler) {
-        this.networkScannerService = networkScannerService;
-        this.networkScanScheduler = networkScanScheduler;
+    public ScanController(ScanService scanService,
+                          ScanScheduler scanScheduler) {
+        this.scanService = scanService;
+        this.scanScheduler = scanScheduler;
     }
 
     public static class IntervalRequest {
@@ -33,22 +33,22 @@ public class ScanController {
 
     @GetMapping("/enabled")
     public ResponseEntity<Boolean> checkScanEnabled() {
-        return ResponseEntity.ok(networkScannerService.isScanEnabled());
+        return ResponseEntity.ok(scanService.isScanEnabled());
     }
 
     @GetMapping("/enable/{enabled}")
     public ResponseEntity<Boolean> enableScan(@PathVariable Boolean enabled) {
-        return ResponseEntity.ok(networkScannerService.setScanEnabled(enabled));
+        return ResponseEntity.ok(scanService.setScanEnabled(enabled));
     }
 
     @GetMapping("/interval")
     public ResponseEntity<Long> getScanInterval() {
-        return ResponseEntity.ok(networkScanScheduler.getScanIntervalMs());
+        return ResponseEntity.ok(scanScheduler.getScanIntervalMs());
     }
 
     @PutMapping("/setinterval")
     public ResponseEntity<Long> setScanInterval(@RequestBody IntervalRequest req) {
-        return ResponseEntity.ok(networkScanScheduler.updateScanInterval(req.getInterval()));
+        return ResponseEntity.ok(scanScheduler.updateScanInterval(req.getInterval()));
     }
 
 }
