+++ backend/src/main/resources/application.properties
@@ -1,3 +1,15 @@
 spring.application.name=LanSentinel
 lansentinel.scan-interval-ms=30000
 lansentinel.subnet-to-scan=192.168.100.0/24
+
+# --- DataSource Configuration ---
+spring.datasource.url=jdbc:postgresql://192.168.100.205:5432/postgres
+spring.datasource.username=postgres
+spring.datasource.password=
+spring.datasource.driver-class-name=org.postgresql.Driver
+
+# --- JPA / Hibernate ---
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=false
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
