+++ frontend/src/app/service/device.service.ts
@@ -7,7 +7,7 @@ import {Observable} from 'rxjs';
   providedIn: 'root'
 })
 export class DeviceService {
-  private apiUrl = 'http://localhost:8080/devices';
+  private apiUrl = 'http://localhost:8080/device';
 
   constructor(private http: HttpClient) {}
